Adding fixture file for bdd
